Version 8.0.
Various minor and more major improvements on the development version over the last year,
which are now ported in this release.

1) A completely new CMakeLists file that also installs various external libraries if they
are required (boost, eigen, etc.) or requested (pastix, Umfpack, etc.) and not found in
the system.

CMakeLists.txt has been replaced by CMakeLists-template.txt to prevent git from clashing with
local, user-specific choices (just copy this template into CMakeLists.txt and modify the later).

The local CMakeLists also include Eigen in a cleaner way, through Eigen::Eigen instead of
TARGER_INCLUDE_DIRECTORIES.

2) A new class, "LinearSolver", can be used in your code to allow you to select the required solver
on the fly. See example of usage in HHO-full-gradient-diff.
HHO-MHD has been adapted to use any solver recognised in the external and not depend solely on MKL.

3) New schemes:
    - DDR scheme for Navier Stokes
    - HHO for linear diffusion, using a new class (HHOCore) which is modeled on DDRCore. The old HHO
    classes HybridCore and HHO-general will not be maintained any longer.

4) Class Manicore to deal with alternating forms.

5) Fixed a bug in creating boundary/internal edges during mesh generation.

6) New class BChandlers with some members to help handle general boundary conditions,
and updated boundarycondition (to compute the number of dirichlet edges and vertices,
and with structures to define planes and convex hulls; see usage in sddr-navier-stokes).

7) local_static_condensation: Changed label "_sys" into "_gl" for globally coupled DOFs.

8) Fix various VectorXd which should have been VectorXi (in variable_dofs_space, serendipity, etc.)

9) Serendipity DDR: option to not use serendipity.

10) sxcurl, sxgrad: member function to compute vertex values based on potential reconstructions.